+++ ci/do_ci.sh
@@ -87,10 +87,10 @@ fi
 CMAKE_OPTIONS+=("-DCMAKE_CXX_STANDARD_REQUIRED=ON")
 CMAKE_OPTIONS+=("-DCMAKE_CXX_EXTENSIONS=OFF")
 
-CMAKE_BUILD_ARGS=(--parallel)
-
-if [[ "${OTELCPP_CMAKE_VERBOSE_BUILD:-OFF}" =~ ^(1|ON|on|TRUE|true|YES|yes)$ ]]; then
-  CMAKE_BUILD_ARGS+=(--verbose)
+if [ -n "${OTELCPP_CMAKE_BUILD_ARGS}" ]; then
+  read -ra CMAKE_BUILD_ARGS <<< "${OTELCPP_CMAKE_BUILD_ARGS}"
+else
+  CMAKE_BUILD_ARGS=(--parallel)
 fi
 
 if command -v ninja >/dev/null 2>&1; then
@@ -621,14 +621,29 @@ elif [[ "$1" == "code.coverage" ]]; then
   cd "${BUILD_DIR}"
   rm -rf *
   cmake "${CMAKE_OPTIONS[@]}"  \
-        -DCMAKE_CXX_FLAGS="-Werror --coverage $CXXFLAGS" \
+        -DCMAKE_CXX_FLAGS="-Werror --coverage -fprofile-update=atomic $CXXFLAGS" \
+        -C "${SRC_DIR}/test_common/cmake/all-options-abiv2-preview.cmake" \
+        -DWITH_EXAMPLES=OFF \
+        -DWITH_EXAMPLES_HTTP=OFF \
+        -DWITH_BENCHMARK=OFF \
         "${SRC_DIR}"
+
   cmake --build . "${CMAKE_BUILD_ARGS[@]}"
   ctest --output-on-failure
-  lcov --directory $PWD --capture --output-file coverage.info
-  # removing test http server coverage from the total coverage. We don't use this server completely.
-  lcov --remove coverage.info '*/ext/http/server/*'> tmp_coverage.info 2>/dev/null
-  cp tmp_coverage.info coverage.info
+
+  lcov --directory "$PWD" --capture \
+  --include "${SRC_DIR}/*" \
+  --exclude "${SRC_DIR}/third_party/*" \
+  --exclude "${SRC_DIR}/test_common/*" \
+  --exclude "${SRC_DIR}/*/test/*" \
+  --exclude "${SRC_DIR}/functional/*" \
+  --exclude "${SRC_DIR}/semconv/*" \
+  --exclude "${SRC_DIR}/examples/*" \
+  --ignore-errors unused \
+  --output-file coverage.info
+
+  echo "Code coverage output file generated at ${BUILD_DIR}/coverage.info. To generate an HTML report run:"
+  echo "genhtml ${BUILD_DIR}/coverage.info -o coverage_html"
   exit 0
 fi